chapter_004/src/main/java/ru/job4j/stream/exercises/CollectToAnyCollection.java
package ru.job4j.stream.exercises;

import java.util.LinkedHashSet;
import java.util.Set;
import java.util.stream.Collectors;
import java.util.stream.Stream;

/**
 * Сборка элементов. Сборка в произвольную реализацию Collection.
 * <p>
 * Создать из стрима коллекцию LinkedHashSet.
 *
 * @author ViktorJava (gipsyscrew@gmail.com)
 * @version 0.1
 * @since 21.12.2020
 */
public class CollectToAnyCollection {
    public static Set<Integer> collect(Stream<Integer> data) {
        return data.collect(Collectors.toCollection(LinkedHashSet::new));
    }
}
